Move resname to base unit
